requirements.txt
Flask
Flask-Login
python-dotenv
PyMySQL
requests
msal
numpy